+++ Source/art/ftfont.m
@@ -113,7 +113,7 @@ static int subpixel_text;
   unsigned int cachedGlyph[CACHE_SIZE];
   NSSize cachedSize[CACHE_SIZE];
 
-  float lineHeight;
+  CGFloat lineHeight;
 }
 @end
 
@@ -167,7 +167,7 @@ static FT_Error ft_get_face(FTC_FaceID fid, FT_Library lib,
 @implementation FTFontInfo
 
 - (id) initWithFontName: (NSString *)name
-                 matrix: (const float *)fmatrix
+                 matrix: (const CGFloat *)fmatrix
              screenFont: (BOOL)p_screenFont
 {
   NSArray *rfi;
@@ -411,7 +411,7 @@ static FT_Error ft_get_face(FTC_FaceID fid, FT_Library lib,
   return coveredCharacterSet;
 }
 
-- (float) defaultLineHeightForFont
+- (CGFloat) defaultLineHeightForFont
 {
   return lineHeight;
 }
@@ -1753,7 +1753,7 @@ static FT_Error ft_get_face(FTC_FaceID fid, FT_Library lib,
 }
 
 
-- (float) widthOfString: (NSString*)string
+- (CGFloat) widthOfString: (NSString*)string
 {
   unichar ch;
   int i, c = [string length];
